requirements.txt
-i https://pypi.org/simple
annotated-doc==0.0.4; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
antlr4-python3-runtime==4.9.3
anyio==4.13.0; python_version >= '3.10'
asn1crypto==1.5.1
asttokens==3.0.1; python_version >= '3.8'
attrs==25.4.0; python_version >= '3.9'
aws-secretsmanager-caching==1.1.3; python_version >= '3.6'
black[jupyter]==26.3.1; python_version >= '3.10'
boto3==1.42.95; python_version >= '3.9'
botocore==1.42.95; python_version >= '3.9'
certifi==2026.4.22; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.7; python_version >= '3.7'
click==8.3.3; python_version >= '3.10'
cloudpickle==3.1.2; python_version >= '3.8'
cryptography==46.0.7; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
decorator==5.2.1; python_version >= '3.8'
dill==0.4.1; python_version >= '3.9'
docker==7.1.0; python_version >= '3.8'
executing==2.2.1; python_version >= '3.8'
fastapi==0.136.1; python_version >= '3.10'
filelock==3.29.0; python_version >= '3.10'
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.8; python_version >= '3.7' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
h11==0.16.0; python_version >= '3.8'
idna==3.13; python_version >= '3.8'
importlib-metadata==6.11.0; python_version >= '3.8'
ipython==9.13.0; python_version >= '3.11'
ipython-pygments-lexers==1.1.1; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jmespath==1.1.0; python_version >= '3.9'
jsonschema==4.26.0; python_version >= '3.10'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
markdown-it-py==4.0.0; python_version >= '3.10'
matplotlib-inline==0.2.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
mock==4.0.3; python_version >= '3.6'
multiprocess==0.70.19; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
numpy==2.4.4; python_version >= '3.11'
omegaconf==2.3.0; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pandas==3.0.2; python_version >= '3.11'
parso==0.8.6; python_version >= '3.6'
pathos==0.3.5; python_version >= '3.9'
pathspec==1.1.0; python_version >= '3.9'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
platformdirs==4.9.6; python_version >= '3.10'
polars==1.40.1; python_version >= '3.10'
polars-runtime-32==1.40.1; python_version >= '3.10'
pox==0.3.7; python_version >= '3.9'
ppft==1.7.8; python_version >= '3.9'
prompt-toolkit==3.0.52; python_version >= '3.8'
protobuf==6.33.6; python_version >= '3.9'
psutil==7.2.2; python_version >= '3.6'
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==3.0; python_version >= '3.10'
pydantic==2.13.3; python_version >= '3.9'
pydantic-core==2.46.3; python_version >= '3.9'
pygments==2.20.0; python_version >= '3.9'
pyjwt==2.12.1; python_version >= '3.9'
pyopenssl==26.0.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.2.2; python_version >= '3.10'
pytokens==0.4.1; python_version >= '3.8'
pytz==2026.1.post1
pyyaml==6.0.3; python_version >= '3.8'
referencing==0.37.0; python_version >= '3.10'
requests==2.33.1; python_version >= '3.10'
rich==14.3.4; python_full_version >= '3.8.0'
rpds-py==0.30.0; python_version >= '3.10'
ruamel.yaml==0.19.1; python_version >= '3.9'
s3transfer==0.16.1; python_version >= '3.9'
sagemaker==2.257.2; python_version >= '3.9'
sagemaker-core==1.0.78; python_version >= '3.8'
schema==0.7.8
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smdebug-rulesconfig==1.0.1; python_version >= '2.7'
snowflake-connector-python==4.4.0; python_version >= '3.9'
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==1.0.0; python_version >= '3.10'
tblib==3.2.2; python_version >= '3.9'
tokenize-rt==6.2.0; python_version >= '3.9'
tomlkit==0.14.0; python_version >= '3.9'
tqdm==4.67.3; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
uvicorn==0.46.0; python_version >= '3.10'
wcwidth==0.6.0; python_version >= '3.8'
zipp==3.23.1; python_version >= '3.9'
